requirements.txt
fastapi==0.68.2
requests==2.26.0
neo4j==4.4.0
nltk==3.6.6
uvicorn==0.15.0
python-dotenv